fix: Make AbstractOperatorExtension use the kubernetesClient if the infrastructureKubernetesClient is not set
