--- operator-framework-junit5/src/test/java/io/javaoperatorsdk/operator/junit/LocallyRunOperatorExtensionIT.java
@@ -1,42 +1,0 @@
-/*
- * Copyright Java Operator SDK Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *         http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.javaoperatorsdk.operator.junit;
-
-import java.nio.file.Path;
-import java.util.List;
-
-import org.junit.jupiter.api.Test;
-
-import io.fabric8.kubernetes.client.KubernetesClientBuilder;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class LocallyRunOperatorExtensionIT {
-
-  @Test
-  void getAdditionalCRDsFromFiles() {
-    System.out.println(Path.of("").toAbsolutePath());
-    System.out.println(Path.of("src/test/crd/test.crd").toAbsolutePath());
-    final var crds =
-        LocallyRunOperatorExtension.getAdditionalCRDsFromFiles(
-            List.of("src/test/resources/crd/test.crd", "src/test/crd/test.crd"),
-            new KubernetesClientBuilder().build());
-    assertNotNull(crds);
-    assertEquals(2, crds.size());
-    assertEquals("src/test/crd/test.crd", crds.get("externals.crd.example"));
-    assertEquals("src/test/resources/crd/test.crd", crds.get("tests.crd.example"));
-  }
-}
